Source code refactoring
