+++ src/algorithms/needleman-wunsch.js
