+++ src/algorithms/smith-waterman.js
